Refactors and typos
